+++ CHANGELOG.md
@@ -12,6 +12,8 @@ BACKWARDS INCOMPATIBILITIES:
 BUG FIXES:
 
   * core: Variables are validated to not contain interpolations. [GH-180]
+  * providers/heroku: If you delete the `config_vars` block, config vars
+      are properly nuked.
 
 PLUGIN CHANGES: